+++ src/main/java/racinggame/util/Utils.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.util;
 
 public class Utils {
