+++ src/test/java/racinggame/UtilsTest.java
@@ -3,6 +3,7 @@ package racinggame;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
+import racinggame.util.Utils;
 
 class UtilsTest {
